requirements.txt
fastapi
uvicorn[standard]
langgraph
langchain-google-genai
pymongo
python-dotenv
pydantic
pytz
google-genai